refac(read files): move functions to new file_io package
