+++ src/careamics/dataset/dataset_utils/iterate_over_files.py
@@ -10,9 +10,9 @@ from torch.utils.data import get_worker_info
 
 from careamics.config import DataConfig, InferenceConfig
 from careamics.utils.logging import get_logger
+from careamics.file_io.read import read_tiff
 
 from .dataset_utils import reshape_array
-from .read_tiff import read_tiff
 
 logger = get_logger(__name__)
