+++ src/careamics/dataset/iterable_pred_dataset.py
@@ -9,10 +9,12 @@ from numpy.typing import NDArray
 from torch.utils.data import IterableDataset
 
 from careamics.transforms import Compose
+from careamics.file_io.read import read_tiff
+
 
 from ..config import InferenceConfig
 from ..config.transformations import NormalizeModel
-from .dataset_utils import iterate_over_files, read_tiff
+from .dataset_utils import iterate_over_files
 
 
 class IterablePredDataset(IterableDataset):
