+++ src/careamics/file_io/__init__.py
@@ -1,0 +1,1 @@
+from . import read
