+++ src/careamics/file_io/read/__init__.py
@@ -1,0 +1,3 @@
+from .get_func import get_read_func
+from .tiff import read_tiff
+from .zarr import read_zarr
