+++ src/careamics/file_io/read/tiff.py
