+++ src/careamics/file_io/read/zarr.py
